requirements.txt
numpy
Pillow
pyx